Add constraint-based type checking
Type checking an expression is now divided into two phases:

1. Constraints are generated for each TypeDeductionVar
2. For each TypeDeductionVar a single type is generated that satisfies
   all the constraints (if that is not possible an error is reported).
   This type is then inserted into the TypeEnv that is used for
   substitution.

This replaces the previous approach where the TypeEnv was updated as new
information became available. For example, a TypeDeductionVar might at
one point be mapped to a $KnownIntType(5, 5) and later to i8. This was
hard to reason about and I'm not completely sure if doing updates like
this would always be correct. With the new approach, we only assign a
type to a TypeDeductionVar once we know *all* the constraints, so no
updates are necessary. This is easier to understand and more flexible.