Fix type deduction for KnownIntTypes
The constraint system I initially implemented was under-constrained when
it came to KnownIntTypes. It still is, but I now added a second step to
manually chose the best integer type for known ints (for my definition
of "best").

For example:

    proc max'T (a: T, b: T) -> T {}

    let s = max(max(1, 2), 220 as u8);

I want the type of `s` to be deduced as `u8`. But in my original purely
bottom-up approach, `max(1, 2)` would be of type `i32`, which resulted
in the whole expression being of type `i32`.

My fix was to add another step that goes top-down, which allows the
propagation of preferred integer types. In the above example, since
`220` is of type `u8`, this information is propagated down to
`max(1, 2)`, ensuring that it also gets the type `u8`.